+++ Kobold2D-2.0.3/HoopyDrop/Projectfiles/PhysicsLayer.h
@@ -35,6 +35,6 @@ enum
 -(void) removeYellowThing: (CCSprite*) sprite;
 -(void) removeGreenThing: (CCSprite*) sprite;
 -(void) removePurpleThing: (CCSprite*) sprite;
--(void) addTarget:(CollisionHandler*) handler andBaseSprite: (NSString*)baseSpriteName andParentNode: (int) parentNodeTag andTrackedBy: (NSMutableArray*) trackingArray at: (double) createTime;
+-(void) addTarget:(CollisionHandler*) handler andBaseSprite: (NSString*)baseSpriteName andParentNode: (int) parentNodeTag andTrackedBy: (NSMutableArray*) trackingArray at: (uint) createTime;
 
 @end
